게시물 삭제시 확인절차 구현

+++ src/main/webapp/jsp/article/detail.jsp
@@ -22,7 +22,7 @@ Map<String, Object> articleRow = (Map<String, Object>)request.getAttribute("arti
 	<div>
 		<a href="list">목록</a>
 		<a href="modify?id=<%= (int)articleRow.get("id") %>">수정</a>
-		<a href="doDelete?id=<%= (int)articleRow.get("id") %>">삭제</a>
+		<a href="doDelete?id=<%= (int)articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
 	</div>
 </body>
 </html>
